+++ springjdbc/src/main/java/StudyJDBC/springjdbc/repository/ArticleRepository.java
@@ -5,8 +5,8 @@ import java.util.List;
 import StudyJDBC.springjdbc.domain.Article;
 public interface ArticleRepository {
     List<Article> findAll();
-    Article findById(int id);
+    Article findById(Long id);
     Article insert(Article article);
-    Article update(int id, Article article);
-    void deleteById(int id);
+    Article update(Long id, Article article);
+    void deleteById(Long id);
 }
