+++ springjdbc/src/main/java/StudyJDBC/springjdbc/repository/BoardRepository.java
@@ -5,6 +5,6 @@ import StudyJDBC.springjdbc.domain.Board;
 
 public interface BoardRepository {
     List<Board> findAll();
-    Board findById(int id);
+    Board findById(Long id);
     void insert(Board board);
 }
