+++ springjdbc/src/main/java/StudyJDBC/springjdbc/repository/MemberRepository.java
@@ -5,6 +5,6 @@ import StudyJDBC.springjdbc.domain.Member;
 
 public interface MemberRepository {
     List<Member> findAll();
-    Member findById(int id);
+    Member findById(Long id);
     void insert(Member member);
 }
